+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Add signature to model
 - Improve immutability
+- Improve readability of visibility cache
 
 # Version 2025.8.8 (2025-08-23)