+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 2025.8.9 (2025-08-23)
 
 - Add signature to model
+- Improve decorators
 - Improve immutability
 - Improve readability of visibility cache